[#957] Add IntlWrapper component as we need to rewrap Intl for modals
