+++ client/src/containers/App.jsx
@@ -1,28 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Router, Route, IndexRedirect } from 'react-router';
-import { addLocaleData, IntlProvider } from 'react-intl';
-import { connect } from 'react-redux';
-import fr from 'react-intl/locale-data/fr';
-import en from 'react-intl/locale-data/en';
+import IntlWrapper from './IntlWrapper';
 import Library from '../components/Library';
 import Visualisation from './Visualisation';
 import Dataset from './Dataset';
 import Dashboard from './Dashboard';
 import Users from '../components/Users';
 import Main from './Main';
-import enTranslations from '../translations/en.json';
-import frTranslations from '../translations/fr.json';
 
-addLocaleData(fr, en);
-
-function App({ history, translations }) {
-  const locale = translations.language;
-
-  const messages = locale === 'en' ? enTranslations : frTranslations;
-
+export default function App({ history }) {
   return (
-    <IntlProvider locale={locale} messages={messages}>
+    <IntlWrapper>
       <Router history={history}>
         <Route path="/" component={Main}>
           <IndexRedirect from="" to="library" />
@@ -36,20 +25,10 @@ function App({ history, translations }) {
           <Route path="admin/users" component={Users} />
         </Route>
       </Router>
-    </IntlProvider>
+    </IntlWrapper>
   );
 }
 
 App.propTypes = {
   history: PropTypes.object.isRequired,
-  translations: PropTypes.any,
 };
-
-
-function mapStateToProps(state) {
-  return {
-    translations: state.translations,
-  };
-}
-
-export default connect(mapStateToProps)(App);
