+++ client/src/containers/Main.jsx
@@ -32,6 +32,7 @@ Main.propTypes = {
 function mapStateToProps(state) {
   return {
     notification: state.notification,
+    modalVisible: state.activeModal != null,
   };
 }
